requirements.txt
streamlit==1.32.2
psycopg2-binary==2.9.9
fpdf2==2.7.7
pandas==2.1.4
python-dotenv==1.0.0